Dalszy ciąg zadania

+++ src/main/java/pl/sdacademy/animals/Animal.java
@@ -4,5 +4,5 @@ public interface Animal {
 
     boolean isAlive();
 
-    int getWeight();
+    double getWeight();
 }
